move common cde into  common module and use ssh-config command to discover hostname for winrm
